GSF: Automatically handle cleanup of files deleted outside the IDE.
During startup, for each index, I already have a list of all the files known to the index.
(I load a special map of url to timestamp for each such file by using Lucene's support
to only load specified terms; this map is used during startup indexing to quicklycheck whether a file is new or has been updated outside of the IDE.).
Now, I also keep a list of all the files I've indexed that I also had a timestamp for (e.g. it
was known to the index). After indexing each source root, I check whether the size of these
two lists are identical.  If not, files have been deleted outside of the IDE, and I diff
the two lists and delete the files that were in the index but no longer on disk.

(This was requested by PHP but is useful for all language clients since you can now be confident
that all matches found in the index are correct.)