148504:  Javascript is not parsed after some changes in a php file
The incremental model updates need to not only look at the edited ranges, but also check the token hierarchy such
that if you for example add a new <script>foo</script> string in your document, this is seen as a new embedded
JavaScript section that must be parsed and processed. The EditHistory now listens on the TokenHierarchy, and
all the incremental models that need it now check for relevant token ids. For example, the JavaScript embedding
model looks for <script> or JavaScript attributes (onclick, onsubmit etc), the Ruby and YAML models look for
added or removed <% delimiters, the CSS model looks for <style> or html attribute value changes, and so on.

Unfortunately, the TokenHierarchyListener doesn't quite work with embedded regions yet, so for now, in JSP, PHP
and ERB files, any edit in the HTML section will force the models to recompute. After checking in this code I'm
hoping to get some help debugging this such that hopefully only specific <script> edits (for example) will
cause a model recomputation.

While fixing this, I ran into a problem with the EditHistory, where it was possible for the size delta (which is used
to update code blocks incrementally) to be wrong. This is also fixed along with unit tests to check for it.