Auto-detect python 3.5 and above on windows
Starting with Python 3.5, Python is installed in C:\Program Files\Python35. The
current implementation of PythonPlatformManager.findPlatformProperties doesn't
support spaces. I made a patch to correct this. I tested auto-detect on Windows
and on Linux and all seems fine. But if I remember correctly, path in commands
may cause problems on Windows so I'd like input on this. The patch is attached
to this project.